restful-api-server/requirements.txt
# FastAPI 핵심 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스 관련
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# JWT 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 환경 변수 및 설정
pydantic-settings==2.0.3
python-dotenv==1.0.0

# 파일 업로드
aiofiles==23.2.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2